Add 0x72/0x73 Jcc instructions, modify Jcc instructions to handle relatives correctly, modify file reading

+++ src/main/kotlin/org/bread_experts_group/breadmod/experimental/computer/disc/iso9960/DirectoryRecord.kt
@@ -41,8 +41,8 @@ class DirectoryRecord(
 			val length = stream.read()
 			/*val extendedAttributeRecordLength = */stream.read()
 			val record = DirectoryRecord(
-				stream.readLSBMSB(4),
-				stream.readLSBMSB(4),
+				stream.readLSBMSB(4).toInt(),
+				stream.readLSBMSB(4).toInt(),
 				ZonedDateTime.of(
 					stream.read() + 1900,
 					stream.read(),
