+++ src/main/kotlin/org/bread_experts_group/breadmod/experimental/computer/ia32/instruction/H74InstructionJE.kt
@@ -6,7 +6,7 @@ import org.bread_experts_group.breadmod.experimental.computer.ia32.IA32Processor
 object H74InstructionJE {
 	fun handle(processor: IA32Processor) {
 		if (processor.csOverride || processor.bitOverride) TODO("can't support CS/66")
-		val relative = processor.fetch().let { processor.cir.toInt() }
+		val relative = processor.fetch().let { processor.cir.toByte() }
 		if (processor.getFlag(FlagType.ZERO_FLAG)) {
 			processor.logger.warn("JE SHORT $relative")
 			processor.ip.ex = (processor.ip.ex.toInt() + relative).toULong()
